backend/requirements.txt
python-dotenv
openai
fastapi
uvicorn
pydantic
mangum